requirements.txt
flask
python-dotenv
openai
aiohttp
requests
gunicorn
google-cloud-firestore
Werkzeug
PyJWT